libobs

Function obs_properties_add_path

pub unsafe extern "C" fn obs_properties_add_path(
    props: *mut obs_properties_t,
    name: *const c_char,
    description: *const c_char,
    type_: obs_path_type,
    filter: *const c_char,
    default_path: *const c_char,
) -> *mut obs_property_t

Adds a ‘path’ property. Can be a directory or a file.

If target is a file path, the filters should be this format, separated by double semicolons, and extensions separated by space: “Example types 1 and 2 (*.ex1 .ex2);;Example type 3 (.ex3)”

@param props Properties object @param name Settings name @param description Description (display name) of the property @param type Type of path (directory or file) @param filter If type is a file path, then describes the file filter that the user can browse. Items are separated via double semicolons. If multiple file types in a filter, separate with space.