libobs

Function obs_reset_video

pub unsafe extern "C" fn obs_reset_video(
    ovi: *mut obs_video_info,
) -> c_int

Sets base video output base resolution/fps/format.

@note This data cannot be changed if an output is currently active. @note The graphics module cannot be changed without fully destroying the OBS context.

@param ovi Pointer to an obs_video_info structure containing the specification of the graphics subsystem, @return OBS_VIDEO_SUCCESS if successful OBS_VIDEO_NOT_SUPPORTED if the adapter lacks capabilities OBS_VIDEO_INVALID_PARAM if a parameter is invalid OBS_VIDEO_CURRENTLY_ACTIVE if video is currently active OBS_VIDEO_MODULE_NOT_FOUND if the graphics module is not found OBS_VIDEO_FAIL for generic failure